puffer.js: ack the last fragment only after chunk is appended to SourceBuffer
Otherwise, the video/audio playback buffer carried on the ACK will
still be the old value, which will actually be updated when the chunk is
appended to SourceBuffer successfully, i.e., after "updateend" event is fired.
An updated buffer level in the ACK helps the server maintain client's
buffer level correctly.