Stop checking if video resolution is higher than client's screen size
screen.width and screen.height are not actually the real physical resolution;
they are in the unit of CSS pixels, and can be affected by the page zoom
level (e.g., in Firefox), window.devicePixelRatio (on mobile devices),
etc.

It is almost impossible to obtain the correct physical resolution, and thus
impossible to reliably check if a video format is an "overkill", i.e., has a
higher resolution than the screen.